Starting to style app
Now you can download by pressing a button (more consistent aesthetic).

Have made the image and the controls on the same row when the app is wide,
different rows when small. It'd be nice if the controls would be aligned to the
bottom or the middle of the image part.