dcc.Graph edits correct shape now
The problem was that @nicholas-esterer originally filtered the shapes in the
table into two sets: those that were new and those than were old. These sets
were represented as lists and were concatenated old + new to get the new list of
shapes. This made it so that the order of the shapes in the dcc.Graph was not
the same as those shown in the data table, so the next time a shape was edited,
it edited the wrong entry in the table.

To fix this the table is just kept as a single list and the indices of the new
entries are found and have a timestamp added to them, then the indices of the
old entries are found and the old timestamp is looked up in the store (by
searching for the matching shape in the store as we cannot simply index into the
store because the list of shapes in the datatable might be different than the
store, say if a shape was deleted).